+++ src/StateMachine/State/RecruitedState.cs
@@ -14,6 +14,7 @@ using NpcAdventure.AI;
 using NpcAdventure.Events;
 using NpcAdventure.Dialogues;
 using NpcAdventure.Utils;
+using StardewValley.Monsters;
 
 namespace NpcAdventure.StateMachine.State
 {
@@ -159,7 +160,7 @@ namespace NpcAdventure.StateMachine.State
             // Fix spawn ladder if area is infested and all monsters is killed but NPC following us
             if (this.StateMachine.Companion.currentLocation is MineShaft mines && mines.mustKillAllMonstersToAdvance())
             {
-                var monsters = from c in mines.characters where c.IsMonster select c;
+                var monsters = from c in mines.characters where c.IsMonster && !this.ai.IsLovedMonster(c as Monster) select c;
                 if (monsters.Count() == 0)
                 {
                     Vector2 vector2 = this.StateMachine.Reflection.GetProperty<Vector2>(mines, "tileBeneathLadder").GetValue();